requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0
aiosqlite>=0.19.0
flask>=3.0.0
flask-cors>=4.0.0
Flask-Limiter>=3.5.0
waitress>=3.0.0
cachetools>=5.3.0
thefuzz>=0.22.0
better-profanity>=0.7.0
